Break out the simple cases
